+++ src/Khatovar/Bundle/ExactionBundle/DependencyInjection/KhatovarExactionExtension.php
@@ -41,6 +41,7 @@ class KhatovarExactionExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('form_subscribers.yml');
         $loader->load('form_types.yml');
         $loader->load('managers.yml');
         $loader->load('twig.yml');